requirements.txt
asgiref==3.2.7
Django==3.0.5
django-crispy-forms==1.9.0
django-extensions==2.2.9
django-rest-framework==0.1.0
djangorestframework==3.11.0
pkg-resources==0.0.0
pytz==2020.1
six==1.14.0
sqlparse==0.3.1
